+++ DatabaseApi/Models/MappableModel.cs
@@ -1,6 +1,5 @@
 // System
 using System;
-using System.Numerics;
 
 
 namespace DatabaseApi.Models
@@ -9,7 +8,7 @@ namespace DatabaseApi.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public BigInteger? WebId { get; set; }
+        public long? WebId { get; set; }
         public string WebName { get; set; }
         public int? TargetWeb { get; set; }
         public DateTime? LastUpdate { get; set; }
